Significant changes
  - Added the 'blocks/' folder to hold gw components.  I've
    moved some of Mary Lou's blocks into this folder and this
    weekend I really need to clean up Alfreds blocks.  I want a
    cleaner library for the overall Arduino environment and this
    is a key to getting that.

+++ messages/state_vec.h
@@ -18,10 +18,10 @@ struct State_vec_msg : public Message {
 	//      const char* name()
 	//      int id()
 	
-	int theta;
-	int theta_dot;
-	int x;
-	int x_dot;
+	double theta;
+	double theta_dot;
+	double x;
+	double x_dot;
     
     //minimal constructor
     State_vec_msg() : Message("state_vec"),
